Moved tests and MetricProvider implementations to the expectation file. Added library metadata.
