+++ contrib/experimental/glam_expectations/expectations/expect_column_values_to_be_edtf_parseable.py
@@ -1,4 +1,4 @@
-from datetime import datetime
+from edtf import parse_edtf
 from typing import Dict, List, Optional, Union
 
 import numpy as np
@@ -10,34 +10,47 @@ from great_expectations.execution_engine import (
     PandasExecutionEngine,
     SparkDFExecutionEngine,
 )
+from great_expectations.expectations.expectation import (
+    ColumnMapExpectation,
+    Expectation,
+    ExpectationConfiguration,
+)
+from great_expectations.expectations.metrics.map_metric import (
+    ColumnMapMetricProvider,
+    column_condition_partial,
+)
 from great_expectations.expectations.util import render_evaluation_parameter_string
 
-from ...core.batch import Batch
-from ...data_asset.util import parse_result_format
-from ...execution_engine.sqlalchemy_execution_engine import SqlAlchemyExecutionEngine
-from ...render.renderer.renderer import renderer
-from ...render.types import RenderedStringTemplateContent
-from ...render.util import (
+from great_expectations.render.renderer.renderer import renderer
+from great_expectations.render.types import RenderedStringTemplateContent
+from great_expectations.render.util import (
     num_to_str,
     parse_row_condition_string_pandas_engine,
     substitute_none_for_missing,
 )
-from ..expectation import (
-    ColumnMapExpectation,
-    Expectation,
-    InvalidExpectationConfigurationError,
-    _format_map_output,
-)
-from ..registry import extract_metrics, get_metric_kwargs
 
-try:
-    import sqlalchemy as sa
-except ImportError:
-    pass
+class ColumnValuesEdtfParseable(ColumnMapMetricProvider):
+    condition_metric_name = "column_values.edtf_parseable"
 
+    @column_condition_partial(engine=PandasExecutionEngine)
+    def _pandas(cls, column, **kwargs):
+        def is_parseable(val):
+            try:
+                if type(val) != str:
+                    raise TypeError(
+                        "Values passed to expect_column_values_to_be_edtf_parseable must be of type string.\nIf you want to validate a column of dates or timestamps, please call the expectation before converting from string format."
+                    )
+
+                parse_edtf(val)
+                return True
+
+            except (ValueError, OverflowError):
+                return False
+
+        return column.map(is_parseable)
 
 class ExpectColumnValuesToBeEdtfParseable(ColumnMapExpectation):
-    """Expect column entries to be parsable using [Extended Date/Time Format (EDTF) Specification](https://www.loc.gov/standards/datetime/).
+    """Expect column entries to be parsable using the [Extended Date/Time Format (EDTF) specification](https://www.loc.gov/standards/datetime/).
 
     expect_column_values_to_be_edtf_parseable is a \
     :func:`column_map_expectation <great_expectations.execution_engine.execution_engine.MetaExecutionEngine
@@ -73,6 +86,26 @@ class ExpectColumnValuesToBeEdtfParseable(ColumnMapExpectation):
         :ref:`include_config`, :ref:`catch_exceptions`, and :ref:`meta`.
 
     """
+
+    # These examples will be shown in the public gallery, and also executed as unit tests for your Expectation
+    examples = [
+        {
+            "data": {
+                
+            },
+            "tests": [
+            ]
+        }
+    ]
+
+    # This dictionary contains metadata for display in the public gallery
+    library_metadata = {
+        "maturity": "experimental",  # "experimental", "beta", or "production"
+        "tags": ["edtf", "datetime" , "glam"],
+        "contributors": ["@mielvds"],
+        "package": "experimental_expectations",
+        "requirements": [],
+    }
 
     map_metric = "column_values.edtf_parsable"
     success_keys = ("mostly",)
