--- contrib/experimental/glam_expectations/metrics/column_map_metrics/column_values_edtf_parseable.py
@@ -1,31 +1,0 @@
-from edtf import parse_edtf
-
-from great_expectations.execution_engine import PandasExecutionEngine
-from great_expectations.execution_engine.sqlalchemy_execution_engine import (
-    SqlAlchemyExecutionEngine,
-)
-from great_expectations.expectations.metrics.map_metric import (
-    ColumnMapMetricProvider,
-    column_condition_partial,
-)
-
-
-class ColumnValuesEdtfParseable(ColumnMapMetricProvider):
-    condition_metric_name = "column_values.edtf_parseable"
-
-    @column_condition_partial(engine=PandasExecutionEngine)
-    def _pandas(cls, column, **kwargs):
-        def is_parseable(val):
-            try:
-                if type(val) != str:
-                    raise TypeError(
-                        "Values passed to expect_column_values_to_be_edtf_parseable must be of type string.\nIf you want to validate a column of dates or timestamps, please call the expectation before converting from string format."
-                    )
-
-                parse_edtf(val)
-                return True
-
-            except (ValueError, OverflowError):
-                return False
-
-        return column.map(is_parseable)
